split abstract desk item
